requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
dspy-ai>=2.4.0
jinja2>=3.1.2
python-multipart>=0.0.6
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
apscheduler>=3.10.0
pytest>=7.4.0
httpx>=0.25.0
playwright>=1.40.0
pytest-playwright>=0.4.0
tenacity>=8.0.0
